data is now ts (not tsx) again
